Fixed tile extent updates and made changes to map scene hierarchy
-Fixed issue where tile extent was not being udated on camera perspective change
-Moved SpawnOnMap.cs from the ZoomableMap example folder into our script folder
-Made changes to SpawnOnMap.cs to make spawned objects children of the LocationBasedGame prefab (in prep for minimap)
-Moved both cameras out of the Player prefab to the top of the hierarchy (in prep for minimap)
