+++ FinderKeeper/ProjectSettings/EditorBuildSettings.asset
@@ -17,7 +17,4 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/End.unity
     guid: bef7a9b4e6d134c53a7bc03f9d869e7c
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   m_configObjects: {}
